feat(position): replace old Direction with new enum
