+++ src/main/java/chess/domain/board/position/Direction.java
@@ -1,49 +1,26 @@
 package chess.domain.board.position;
 
-import chess.strategy.DirectionChecker;
-
 public enum Direction {
 
-    UP((eastWardDiff, northWardDiff) -> isUpward(northWardDiff) && isVertical(eastWardDiff)),
-    UP_RIGHT((eastWardDiff, northWardDiff) -> isUpward(northWardDiff) && isRight(eastWardDiff)),
-    RIGHT((eastWardDiff, northWardDiff) -> isHorizontal(northWardDiff) && isRight(eastWardDiff)),
-    DOWN_RIGHT((eastWardDiff, northWardDiff) -> isDownward(northWardDiff) && isRight(eastWardDiff)),
-    DOWN((eastWardDiff, northWardDiff) -> isDownward(northWardDiff) && isVertical(eastWardDiff)),
-    DOWN_LEFT((eastWardDiff, northWardDiff) -> isDownward(northWardDiff) && isLeft(eastWardDiff)),
-    LEFT((eastWardDiff, northWardDiff) -> isHorizontal(northWardDiff) && isLeft(eastWardDiff)),
-    UP_LEFT((eastWardDiff, northWardDiff) -> isUpward(northWardDiff) && isLeft(eastWardDiff));
+    RIGHT(0),
+    UP_RIGHT(45),
+    UP(90),
+    UP_LEFT(135),
+    LEFT(180),
+    DOWN_LEFT(-135),
+    DOWN(-90),
+    DOWN_RIGHT(-45);
 
-    private final DirectionChecker directionChecker;
+    private final int angle;
 
-    Direction(DirectionChecker directionChecker) {
-        this.directionChecker = directionChecker;
+    Direction(int angle) {
+        this.angle = angle;
     }
 
-    public boolean checkByPositionDifference(int eastWardDiff, int northWardDiff) {
-        return directionChecker.checkCorrectDirection(eastWardDiff, northWardDiff);
-    }
+    public boolean hasAngleOf(int x, int y) {
+        double theta = Math.atan2(y, x);
+        double targetAngle = Math.toDegrees(theta);
 
-    private static boolean isUpward(Integer northWardDiff) {
-        return northWardDiff > 0;
-    }
-
-    private static boolean isDownward(Integer northWardDiff) {
-        return northWardDiff < 0;
-    }
-
-    private static boolean isHorizontal(Integer northWardDiff) {
-        return northWardDiff == 0;
-    }
-
-    private static boolean isRight(Integer eastWardDiff) {
-        return eastWardDiff > 0;
-    }
-
-    private static boolean isLeft(Integer eastWardDiff) {
-        return eastWardDiff < 0;
-    }
-
-    private static boolean isVertical(Integer eastWardDiff) {
-        return eastWardDiff == 0;
+        return angle == targetAngle;
     }
 }
